feat(signals): add ExtractStoreFeatureOutput type utility
Extracts output type from signalStoreFeature functions for use as input in dependent features. Simplifies splitting stores into multiple features by eliminating need to manually define input interfaces.

Fix #251
